+++ out/definitionProvider/batariBasicDefinitionProvider.js
@@ -1,0 +1,10 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+const definitionProviderBase_1 = require("./definitionProviderBase");
+class BatariBasicDefinitionProvider extends definitionProviderBase_1.DefinitionProviderBase {
+    constructor() {
+        super("batariBasic");
+    }
+}
+exports.BatariBasicDefinitionProvider = BatariBasicDefinitionProvider;
+//# sourceMappingURL=batariBasicDefinitionProvider.js.map