+++ out/definitionProvider/seventyEightHundredBasicDefinitionProvider.js
@@ -1,0 +1,10 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+const definitionProviderBase_1 = require("./definitionProviderBase");
+class SeventyEightHundredBasicDefinitionProvider extends definitionProviderBase_1.DefinitionProviderBase {
+    constructor() {
+        super("7800basic");
+    }
+}
+exports.SeventyEightHundredBasicDefinitionProvider = SeventyEightHundredBasicDefinitionProvider;
+//# sourceMappingURL=seventyEightHundredBasicDefinitionProvider.js.map