+++ out/documentSymbolProvider/batariBasicDocumentSymbolProvider.js
@@ -1,0 +1,10 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+const documentSymbolProviderBase_1 = require("./documentSymbolProviderBase");
+class BatariBasicDocumentSymbolProvider extends documentSymbolProviderBase_1.DocumentSymbolProviderBase {
+    constructor() {
+        super("batariBasic");
+    }
+}
+exports.BatariBasicDocumentSymbolProvider = BatariBasicDocumentSymbolProvider;
+//# sourceMappingURL=batariBasicDocumentSymbolProvider.js.map