+++ out/documentSymbolProvider/seventyEightHundredBasicOutline.js
@@ -1,0 +1,10 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+const documentSymbolProviderBase_1 = require("./documentSymbolProviderBase");
+class SeventyEightHundredBasicDocumentSymbolProvider extends documentSymbolProviderBase_1.DocumentSymbolProviderBase {
+    constructor() {
+        super("7800basic");
+    }
+}
+exports.SeventyEightHundredBasicDocumentSymbolProvider = SeventyEightHundredBasicDocumentSymbolProvider;
+//# sourceMappingURL=seventyEightHundredBasicOutline.js.map