+++ src/definitionProvider/batariBasicDefinitionProvider.ts
@@ -1,0 +1,9 @@
+"use strict";
+import { DefinitionProviderBase } from './definitionProviderBase';
+
+export class BatariBasicDefinitionProvider extends DefinitionProviderBase {
+    
+    constructor() {
+        super("batariBasic");
+    }
+}