+++ src/definitionProvider/seventyEightHundredBasicDefinitionProvider.ts
@@ -1,0 +1,9 @@
+"use strict";
+import { DefinitionProviderBase } from './definitionProviderBase';
+
+export class SeventyEightHundredBasicDefinitionProvider extends DefinitionProviderBase {
+    
+    constructor() {
+        super("7800basic");
+    }
+}