+++ src/documentSymbolProvider/batariBasicDocumentSymbolProvider.ts
@@ -1,0 +1,9 @@
+"use strict";
+import { DocumentSymbolProviderBase } from './documentSymbolProviderBase';
+
+export class BatariBasicDocumentSymbolProvider extends DocumentSymbolProviderBase {
+    
+    constructor() {
+        super("batariBasic");
+    }
+}