+++ src/documentSymbolProvider/seventyEightHundredBasicDocumentSymbolProvider.ts
@@ -1,0 +1,9 @@
+"use strict";
+import { DocumentSymbolProviderBase } from './documentSymbolProviderBase';
+
+export class SeventyEightHundredBasicDocumentSymbolProvider extends DocumentSymbolProviderBase {
+    
+    constructor() {
+        super("7800basic");
+    }
+}