+++ src/outline/batariBasicOutline.ts
@@ -1,9 +1,0 @@
-"use strict";
-import { OutlineBase } from './outlineBase';
-
-export class BatariBasicOutline extends OutlineBase {
-    
-    constructor() {
-        super("batariBasic");
-    }
-}