+++ src/outline/seventyEightHundredBasicOutline.ts
@@ -1,9 +1,0 @@
-"use strict";
-import { OutlineBase } from './outlineBase';
-
-export class SeventyEightHundredBasicOutline extends OutlineBase {
-    
-    constructor() {
-        super("7800basic");
-    }
-}