+++ CHANGELOG.md
@@ -4,7 +4,8 @@ The following enhancements and changes have been made to Atari Dev Studio:
 
 ## 0.5.5
 
-* Added document outline support for batari Basic and 7800basic
+* Added document outline feature for batari Basic and 7800basic
+* Added Go to Definition feature for batari Basic and 7800basic
 
 ## 0.5.4